minor reorg files
